Add Newtonsoft.Json and update NSwag configuration
Added the `Newtonsoft.Json` package version `13.0.4` to the
`Directory.Packages.props` file and referenced it in the
`Clean.Architecture.Sdk.csproj` file. This ensures the project
can utilize the library.

Updated the `nswag.json` configuration to specify `SystemTextJson`
as the JSON library for code generation by adding the `jsonLibrary`
property.